requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
docutils==0.17.1
idna==3.2
lockfile==0.12.2
luigi==3.0.3
praw==7.3.0
prawcore==2.2.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
tenacity==6.3.1
tornado==6.1
update-checker==0.18.0
urllib3==1.26.6
websocket-client==1.1.0
